+++ CHANGELOG.md
@@ -9,7 +9,9 @@
 	- iterative insert size estimation refining for mate pairs
 	- fasta2diverged.py deprecated
     - `-o/--output` now contains also sspace intermediate files
-    - reduction (fasta2homozygous.py) uses LAST instead of BLAT & multi-threaded
+    - reduction (fasta2homozygous.py) uses LAST instead of BLAT
+      - LAST multi-threaded (only in Python 2.7+)
+    - cleaning-up output directory from intermediate files
 
 ####0.10b
 	- `-l/--limit` now takes number of reads limit as fraction of homozygous genome size. To process all reads set -l to 0.